Change wheel speed conversion factor Prior to this commit, we were scaling the wheel speeds by a factor of 128. This commit changes that to a factor of 50, after on-vehicle testing determined that it was a more reasonable scaling factor.

+++ platforms/kia_soul/firmware/steering/include/steering_control.h
@@ -163,7 +163,7 @@ typedef struct
     float commanded_steering_wheel_angle; /* Angle of steering wheel commanded
                                              by controller. */
 
-    int16_t vehicle_speed; /* Current vehicle speed as reported by the vehicle. */
+    float vehicle_speed; /* Current vehicle speed as reported by the vehicle. */
 
     float commanded_steering_wheel_angle_rate; /* Rate of the steering wheel
                                                   angle commanded by controller. */
